refactor(validator/naming): extract naming health-check into naming/src/health
